feat(Post): add new Factory method
